dev/msp: migrate service accounts, IAM to 'iam' module
