+++ dev/managedservicesplatform/internal/stack/cloudrun/BUILD.bazel
@@ -1,4 +1,3 @@
-load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
@@ -27,23 +26,7 @@ go_library(
         "//dev/managedservicesplatform/spec",
         "//lib/errors",
         "//lib/pointers",
-        "@com_github_aws_jsii_runtime_go//:jsii-runtime-go",
-        "@com_github_grafana_regexp//:regexp",
-        "@com_github_hashicorp_terraform_cdk_go_cdktf//:cdktf",
-        "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google//projectiamcustomrole",
-        "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google//projectiammember",
         "@org_golang_x_exp//maps",
         "@org_golang_x_exp//slices",
-    ],
-)
-
-go_test(
-    name = "cloudrun_test",
-    srcs = ["cloudrun_test.go"],
-    embed = [":cloudrun"],
-    deps = [
-        "//lib/pointers",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
     ],
 )
