+++ dev/managedservicesplatform/internal/stack/project/BUILD.bazel
@@ -18,7 +18,5 @@ go_library(
         "@com_github_grafana_regexp//:regexp",
         "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google//project",
         "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google//projectservice",
-        "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google_beta//googleprojectserviceidentity",
-        "@com_github_sourcegraph_managed_services_platform_cdktf_gen_google_beta//provider",
     ],
 )
