+++ dev/managedservicesplatform/managedservicesplatform.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack"
 	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/cloudrun"
+	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/iam"
 	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/options/terraformversion"
 	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/options/tfcbackend"
 	"github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/project"
@@ -97,9 +98,17 @@ func (r *Renderer) RenderEnvironment(
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create project stack")
 	}
+	iamOutput, err := iam.NewStack(stacks, iam.Variables{
+		ProjectID: *projectOutput.Project.ProjectId(),
+		Image:     build.Image,
+		Service:   svc,
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create IAM stack")
+	}
 	if _, err := cloudrun.NewStack(stacks, cloudrun.Variables{
-		ProjectID:             *projectOutput.Project.ProjectId(),
-		CloudRunIdentityEmail: *projectOutput.CloudRunIdentity.Email(),
+		ProjectID:                      *projectOutput.Project.ProjectId(),
+		CloudRunWorkloadServiceAccount: iamOutput.CloudRunWorkloadServiceAccount,
 
 		Service:     svc,
 		Image:       build.Image,
